requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
cytoolz==0.12.2
eth-account==0.10.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.0
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
exceptiongroup==1.2.0
fonttools==4.47.2
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.6
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lru-dict==1.2.0
matplotlib==3.8.2
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.2.1
parsimonious==0.9.0
pillow==10.2.0
pluggy==1.4.0
protobuf==4.25.2
pycryptodome==3.20.0
pyparsing==3.1.1
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.17.1
six==1.16.0
soupsieve==2.5
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
web3==6.14.0
websockets==12.0
yarl==1.9.4
